tray-monitor: finalize icon cleanup
Drop the unused warning.png resource and keep the
tray icon code on the pre-rendered sprite sheet path.

Tighten the tray icon state handling and type usage
around the animation frame index.

+++ core/src/qt-tray-monitor/systemtrayicon.h
@@ -50,7 +50,7 @@ class SystemTrayIcon : public QSystemTrayIcon {
   Q_DISABLE_COPY(SystemTrayIcon);
 
   IconType currentIcon{IconType::Normal};
-  size_t animationFrameIdx{0};
+  int animationFrameIdx{0};
   bool animationRequested{false};
   QIcon normalIcon;
   QIcon errorIcon;
